google-cloud-run/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pdfplumber==0.11.0
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
pillow==10.2.0
pymupdf==1.23.7
pypdfium2==5.0.0
pikepdf==8.15.1
pytesseract==0.3.10
pdf2docx==0.5.6
camelot-py==0.11.0
tabula-py==2.9.2
pandas==2.1.2
numpy==1.26.2
scipy==1.11.4
jpype1==1.5.0
pytest==7.4.3


